pyproject.toml
[tool.poetry]
name = "ragnorag"
version = "0.1.0"
description = ""
authors = ["llermaly <gllermaly@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.44.0"
streamlit-extras = "^0.6.0"
elasticsearch = "^8.17.2"
st-annotated-text = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
